Ready for Physics
I'll figure out a better camera later. Feel free to clean up.

I made a Ball class in golfCourse.h to hold the radius and initial
location of the ball. I think you need to figure out how to keep track
of its current position. Is there a need for a class for the hole/cup?
There is a struct for wall in case there is a collision event or
something. Why can't this be as simple as Unity?